requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
python-multipart==0.0.6
tiktoken==0.5.1
numpy==1.26.1 